+++ optic/plot.py
@@ -30,45 +30,46 @@ from optic.comm.modulation import detector
 from optic.utils import dB2lin
 import warnings
 
-warnings.filterwarnings('ignore', r'All-NaN (slice|axis) encountered')
+warnings.filterwarnings("ignore", r"All-NaN (slice|axis) encountered")
+
 
 def pconst(x, lim=True, R=1.25, pType="fancy", cmap="turbo", whiteb=True):
     """
     Plot signal constellations.
-    
+
     Parameters
     ----------
     x : complex signals or list of complex signals
         Input signals.
-    
+
     lim : bool, optional
-        Flag indicating whether to limit the axes to the radius of the signal. 
+        Flag indicating whether to limit the axes to the radius of the signal.
         Defaults to True.
-    
+
     R : float, optional
-        Scaling factor for the radius of the signal. 
+        Scaling factor for the radius of the signal.
         Defaults to 1.25.
-    
+
     pType : str, optional
         Type of plot. "fancy" for scatter_density plot, "fast" for fast plot.
         Defaults to "fancy".
-    
+
     cmap : str, optional
         Color map for scatter_density plot.
         Defaults to "turbo".
-    
+
     whiteb : bool, optional
         Flag indicating whether to use white background for scatter_density plot.
         Defaults to True.
-    
+
     Returns
     -------
     fig : Figure
         Figure object.
-    
+
     ax : Axes or array of Axes
         Axes object(s).
-    
+
     """
     if type(x) == list:
         for ind, _ in enumerate(x):
@@ -104,13 +105,14 @@ def pconst(x, lim=True, R=1.25, pType="fancy", cmap="turbo", whiteb=True):
         fig = plt.figure()
 
         if type(x) == list:
-            for k in range(nSubPts):           
-
+            for k in range(nSubPts):
                 for ind in range(len(x)):
                     if pType == "fancy":
                         if ind == 0:
-                            ax = fig.add_subplot(nRows, nCols, Position[k], projection='scatter_density')
-                        ax = constHist(x[ind][:, k], ax, radius, cmap, whiteb)
+                            ax = fig.add_subplot(
+                                nRows, nCols, Position[k], projection="scatter_density"
+                            )
+                        ax = constHist(x[ind][:, k], ax, cmap, whiteb)
                     elif pType == "fast":
                         if ind == 0:
                             ax = fig.add_subplot(nRows, nCols, Position[k])
@@ -126,10 +128,12 @@ def pconst(x, lim=True, R=1.25, pType="fancy", cmap="turbo", whiteb=True):
                     ax.set_xlim(-radius, radius)
                     ax.set_ylim(-radius, radius)
         else:
-            for k in range(nSubPts):                
+            for k in range(nSubPts):
                 if pType == "fancy":
-                    ax = fig.add_subplot(nRows, nCols, Position[k], projection='scatter_density')
-                    ax = constHist(x[:, k], ax, radius, cmap, whiteb)
+                    ax = fig.add_subplot(
+                        nRows, nCols, Position[k], projection="scatter_density"
+                    )
+                    ax = constHist(x[:, k], ax, cmap, whiteb)
                 elif pType == "fast":
[…]
