### 2024-03-04 f324d31
add function to plot colored constellation scatter plots
