### 2024-03-05 d2ef79e
simplify parameter parsing
