### 2024-03-05 559ea93
fix bug by removing njit
