+++ optic/comm/ofdm.py
@@ -74,9 +74,9 @@ def calcSymbolRate(M, Rb, Nfft, Np, G, hermitSym):
     return Rb / (nDataSymbols/(Nfft + G) * np.log2(M))
 
 
-def modulateOFDM(symbTx, param):
+def modulateOFDM(symb, param):
     """
-    OFDM symbols modulator.
+    Modulate OFDM signal.
 
     Parameters
     ----------
@@ -88,7 +88,7 @@ def modulateOFDM(symbTx, param):
                     pilot symbol
     pilotCarriers : np.array
                     indexes of pilot subcarriers
-    symbTx        : complex-valued array
+    symb          : complex-valued array of modulation symbols
                     symbols sequency transmitted
     hermitSym     : boolean
                     True-> Real OFDM symbols / False: Complex OFDM symbols 
@@ -97,7 +97,7 @@ def modulateOFDM(symbTx, param):
     
     Returns
     -------
-    symbTx_OFDM   : complex-valued np.array
+    symb_OFDM   : complex-valued np.array
                     OFDM symbols sequency transmitted
     """
     # Check and set default values for input parameters
@@ -113,22 +113,22 @@ def modulateOFDM(symbTx, param):
 
     # Number of subcarriers
     Ns = Nfft//2 - 1 if hermitSymmetry else Nfft
-    numSymb  = len(symbTx)
+    numSymb  = len(symb)
     numOFDMframes = numSymb//(Ns - Np)
 
     Carriers = np.arange(0, Ns)
     dataCarriers  = np.array(list(set(Carriers) - set(pilotCarriers)))
 
     # Serial to parallel
-    symbTx_par = np.reshape(symbTx, (numOFDMframes, Ns - Np))   
-    symbTx_OFDM_par = np.zeros( (numOFDMframes, SpS*(Nfft + G)), dtype=np.complex64)
+    symb_par = np.reshape(symb, (numOFDMframes, Ns - Np))   
+    sigOFDM_par = np.zeros( (numOFDMframes, SpS*(Nfft + G)), dtype=np.complex64)
     
     for indFrame in range(numOFDMframes):
         # Start OFDM frame with zeros
         frameOFDM = np.zeros(Ns, dtype=np.complex64) 
 
         # Insert data and pilot subcarriers                    
-        frameOFDM[dataCarriers]  = symbTx_par[indFrame, :]
+        frameOFDM[dataCarriers]  = symb_par[indFrame, :]
         frameOFDM[pilotCarriers] = pilot
 
         # Hermitian symmetry
@@ -136,17 +136,18 @@ def modulateOFDM(symbTx, param):
             frameOFDM = hermit(frameOFDM)
        
         # IFFT operation       
-        symbTx_OFDM_par[indFrame, SpS*G : SpS*(G + Nfft)] = ifft(fftshift(padOFDM(frameOFDM, Nfft*(SpS-1)//2))) * np.sqrt(SpS*Nfft)
+        sigOFDM_par[indFrame, SpS*G : SpS*(G + Nfft)] = ifft(fftshift(padOFDM(frameOFDM, (Nfft*(SpS-1))//2))) * np.sqrt(SpS*Nfft)
         
         # Cyclic prefix addition
-        symbTx_OFDM_par[indFrame, 0 : SpS*G] = symbTx_OFDM_par[indFrame, Nfft*SpS : SpS*(Nfft + G)].copy()
+        if G > 0:
+            sigOFDM_par[indFrame, 0 : SpS*G] = sigOFDM_par[indFrame, Nfft*SpS : SpS*(Nfft + G)].copy()
 
-    return symbTx_OFDM_par.ravel()
[…]
