### 2024-04-03 #6400 @ebezzam
hey @lucydot fiber optics is out of my area of expertise so I wouldn't be able to provide critical feedback there. But the digital communications and DSP aspects are more within my domain of knowledge (and also testing Python software). If you think that's enough, I could review
