### 2024-04-10 #6400 @lucydot
Excellent, @ebezzam @taladjidi  - I have added you as reviewers, thank you for agreeing to review this submission to JOSS ⭐

If you haven't reviewed for JOSS before, you can find some detailed information about the review process on the documentation pages: [reviewing for JOSS](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html). You don't need to read this all, but it gives some context if wanted.

The key things that are different from standard journals are: i) everything happens in a Github review thread; ii) the review is a conversation back and forth - you do not need to do you review in a single step; iii) the review is structured in that you work your way through a checklist.

We ask that reviews are completed in about 4 weeks. We advise you start the review early, as it is an iterative process between reviewers and authors.

I will now ask editorialbot to generate a new issue thread which is where the review will take place (this is the pre-review thread). Instructions for you (to generate your review checklist) will be provided there.

If you have any other questions, do not hesitate to contact me. Discussion on the review thread is the best place to contact me, but you can also contact me via email ([l.whalley@northumbria.ac.uk](mailto:l.whalley@northumbria.ac.uk)) if there is something you'd rather not be public.
