### 2024-04-10 #6600 @editorialbot
[REVIEW]: OptiCommPy: Open-source Simulation of Fiber Optic Communications with Python
**Submitting author:** @edsonportosilva (<a href="http://orcid.org/0000-0003-4230-9121">Edson Porto da Silva</a>)
**Repository:** https://github.com/edsonportosilva/OptiCommPy/
**Branch with paper.md** (empty if default branch): add-paper
**Version:** v0.9.0-alpha
