### 2024-04-16 #6600 @ebezzam
@lucydot, @edsonportosilva below is my initial review. I realize that there is an [examples folder](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples), but is it expected to run everything or is there a notebook/script that gives an overview of the main features?

### Functionality

- **Installation**: I would recommend putting the same installation in your README and in the documentation. You can create a `README.rst` file, and use that within your documentation (like [this](https://github.com/ebezzam/python-dev-tips/blob/a76a766f506d871ad8e3a9f396ebb5bdc268addc/docs/source/index.rst?plain=1#L10) instead of creating a [separate file](https://github.com/edsonportosilva/OptiCommPy/blob/main/docs/source/usage.rst?plain=1) so you don’t have to edit twice). I say this because I looked at the [documentation](https://opticommpy.readthedocs.io/en/latest/usage.html) for the installation steps; and `python setup.py install` didn’t work for me, but `pip install .` did (which was in [your README](https://github.com/edsonportosilva/OptiCommPy/blob/452409d2ca8b604ef04aaffe675640eac21c628b/README.md?plain=1#L75)).
- **Functionality**: it would be useful to have a section or notebook to show the [different features](https://opticommpy.readthedocs.io/en/latest/welcome.html), as I’m not sure where I would start to try things out. I realize there’s an [examples folder](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples), but it is not clear where to start.
- **Performance**: it would be interesting to quantify the speed-up thanks to GPU acceleration, as mentioned in Lines 61-64 of the paper.

### Documentation

- **A statement of need**: there is none in the [documentation](https://opticommpy.readthedocs.io/en/latest/index.html).
- **Automated tests**: there are none nor manual steps to verify functionality.
- **Community guidelines**: there is in the [README](https://github.com/edsonportosilva/OptiCommPy?tab=readme-ov-file#how-can-i-contribute), but as mentioned above for “Installation”, it could be nice to use the README in your documentation, as I didn’t see the contribution guidelines at first.

### Software paper

- **A statement of need**: The “Statement of need” starts great by giving the relevant context, but I feel like there’s a paragraph missing after Line 40 to explicitly say something like “OptiCommPy addresses this need by…” and who the target audience is (similar to what’s said in  “Summary”).
- **Quality of writing**: I think Lines 41 and onward should be in a separate section, e.g. with the title “OptiCommPy Code Structure”. And another section on what example scripts/code for the different features could be useful.

### Miscellaneous

- Adding [.ipynb_checkpoints](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples/.ipynb_checkpoints) and [.pylint.d](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples/.pylint.d) to your `.gitignore`
