### 2024-04-18 #6600 @lucydot
Thank you for your thoughtful review @ebezzam

> I realize that there is an [examples folder](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples), but is it expected to run everything or is there a notebook/script that gives an overview of the main features?

With the immediate review process in mind, this seems like an important thing to address. Could you indicate which examples demonstrate the key software functionality @edsonportosilva? This could be useful for future users, so I suggest it is also included in the documentation.
