### 2024-04-18 #6600 @edsonportosilva
Hello @lucydot @ebezzam,

Thanks for the review and the feedback!  In the examples folder, we provide notebooks with simulation setups with various levels of complexity to help the user. The simplest simulation of optical transmission (kind of a 'textbook example')  is provided in [basic_OOK_transmission.ipynb](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/basic_OOK_transmission.ipynb). This notebook would be the starting point e.g. for undergrad students who are taking a first course in optical communications.

The notebook [test_WDM_transmission.ipynb](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/test_WDM_transmission.ipynb) is a more advanced example that deals with coherent optical transmission and uses main features of OptiCommPy (coherent communications, advanced fiber optic modeling, and digital signal processing (DSP)). It is also the kind of simulation targeted by experienced users, such as researchers/engineers working with optical communications. This notebook also automatically calls for GPU processing, if a suitable GPU is available. Thus, by simply running it in e.g. Colab with and without a GPU accelerator, the user can estimate how fast a GPU speeds up the simulation (for this kind of simulation, the difference is quite significant, at least 10x faster I would expect).

The other notebooks are mostly simulation scenarios to "test" various features of the code. In particular, we have included the notebook [test_metrics.ipynb](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/test_metrics.ipynb) where all relevant functions that calculate performance metrics are tested and the results compared with results from communications theory. We are working on setting up proper automated tests for the code in the repository, but it is working in progress.

@ebezzam, I will proceed in fixing the issues listed in your preliminary review.
