### 2024-04-22 #6600 @lucydot
@edsonportosilva - thanks for the update.

As @ebezzam has already mentioned, it would be useful to have documentation to show / walk through the different features as it is difficult to know where to start with the examples folder. Often I see this as a (series of) notebooks displayed in ReadTheDocs. For example, a "Getting Started" code walkthrough showing the basic functionality. You could then link to the examples folder for the more advanced topics. This could also go some way to ensuring that the "Testing" JOSS criteria is met:

> Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
