+++ README.rst
@@ -1,6 +1,8 @@
 .. raw:: html
 
    <p align="center">
+   <img src="https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/logo_OptiCommPy.jpg" width="500">
+   </p>
 
 .. raw:: html
 
@@ -13,9 +15,12 @@ educational and research purposes.
 .. raw:: html
 
    <p align="center">
+   <img class="center" src="https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/eyeDisp.gif" width="400">  <img class="center" src="https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/40GOOK_spectrum.jpg" width="400">   
+   </p>
 
 .. raw:: html
-
+   <p align="center">
+   <img src="https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/DSP.jpg" width="800">
    </p>
 
 |Binder| |Documentation Status| |DOI|
