+++ README.rst
@@ -4,10 +4,6 @@
    <img src="https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/logo_OptiCommPy.jpg" width="500">
    </p>
 
-.. raw:: html
-
-   </p>
-
 This repository is a Python-based framework to simulate systems,
 subsystems, and components of fiber optic communication systems, for
 educational and research purposes.
