+++ README.rst
@@ -14,7 +14,9 @@ educational and research purposes.
    <img class="center" src="https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/eyeDisp.gif" width="400">  <img class="center" src="https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/40GOOK_spectrum.jpg" width="400">   
    </p>
 
+
 .. raw:: html
+   
    <p align="center">
    <img src="https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/DSP.jpg" width="800">
    </p>
