### 2024-04-23 #6600 @taladjidi
@lucydot  @ebezzam

Sorry for the late reply, I was in leave. Below are my first comments.

# Functionality

- Installation went without trouble, though I only tried through `pip install .`

- Examples: some of the notebooks do not run (while I did not test them all, [`basic_EDFA.ipynb`](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/basic_EDFA.ipynb) seems to use an old package structure and does not run. I ran this first since I use an EDFA in the lab and it was the first example in the directory.

- Performance: I agree with @ebezzam that benchmarks curves would be a welcome addition. This could maybe be one of the tests ?

- GPU models / Performance: I don't know how much performance is critical, but [`modelsGPU.py`](https://github.com/edsonportosilva/OptiCommPy/blob/main/optic/models/modelsGPU.py) could benefit greatly from ad hoc kernels, especially for your split step scheme. See for instance the docs from [Cupy](https://docs.cupy.dev/en/stable/user_guide/kernel.html) or [Numba](https://numba.readthedocs.io/en/stable/cuda/index.html). This will help you limit the number of kernels launched. While this may be not critical for 1D problems involved with fibers, this becomes critical in 2D (which might prove relevant depending on the number of modes involved).
Furthermore,  you will gain at least a factor of 2 performance on your FFT's if you explicitely define and cache the FFT plans (see [`get_fft_plan`](https://docs.cupy.dev/en/stable/reference/generated/cupyx.scipy.fftpack.get_fft_plan.html#cupyx.scipy.fftpack.get_fft_plan) or directly through `cupy.cuda.cufft`) instead of using the `cupyx.scipy.fft` API.
- Performance / FFT: I would advise switching to a more performant FFT library than Numpy's since most of you DSP will rely on it. There are a lot of potential candidates, pyFFTW could be a drop in replacement. In the same vein, a lot of operations could be parallelized and JIT compiled using Numba.

# Documentation

- Examples: Jupyter notebooks are a great way to introduce the various functionalities of your software. I found that the graphical outputs were very clear. Maybe it would be good to add Markdown blocks in between the code blocks to link the code with the physical models in use (how does one equation translate in terms of your package functionalities).
- Physical models: I would expect the docstrings of the models to cite some relevant paper (when needed) explaining the model more in detail, or one or two latex equations to succintly define the used conventions etc ...
- In the paper, you compare to commercial / proprietary solutions, maybe you could include a section of your Readme to explain what funcitonalities are supported, or some kind of migration guide ? In a similar fashion, maybe you could include some performance comparisons.

# Software paper

- The manuscript overall is well written and has a very good introduction.
- It would be better to split the Statement of need and the functionalities that you introduce, in order to highlight the problem on one side and its solution on the other.
- The package structure is maybe superfluous in the paper, and maybe could be replaced with benchmarks / comparisons with proprietary software (if you have access to it) ?

Overall it's a very nice project, while the examples need a bit of documentation work, they do seem to cover most of the functionalities of the package.

Best,
