### 2024-04-23 #6600 @klb2
@lucydot @edsonportosilva

I have now completed my review. I tried to limit my comments to new ones that have not been covered in the reviews above.

In general, the package is interesting and contains a lot of features. The main issues are related to the tests and documentation, which lacks significant information (even missing whole modules).

## Documentation
- It would be nice to add a simple example after the installation instructions to verify that the installation worked.
- In general, the documentation does not show any usage examples.
- Required version of matplotlib in documentation (>=1.4.3) does not match that in requirements/setup file (>=3.7.0)
- Pandas is not stated as a dependency in documentation
- It should be mentioned which scripts need to be run to reproduce the figures from the paper.
- The documentation of the `fec` module is missing in the documentation.

## Tests
- `tests/test_dsp.py` does not run with the `unittest` package
- The setup script (`setup.py`) says that the `nose` module is required for testing, but it is not actually required to run the unit tests in `tests/`
- `examples/test_*.py` can not be run with Python (ipython is required). However, this is not mentioned anywhere.
- There only seems to be a small number of unit tests. The code coverage is very low.
- The eye diagrams in `examples/test_dsp_core_functions.ipynb` are not useful (just a blue box).
- Error when running `ipython examples/test_clockRecovery.py`: `ValueError: operands could not be broadcast together with shapes (44000,) (43968,)` (Line 121)
- Error when running `ipython examples/test_fec.py`: `ModuleNotFoundError: No module named 'optic.modulation'` (I think, it should be `optic.comm.modulation`)
- Also, the `optic.fec` module does not exist. I could not find any LDPC functionality in the documentation.
- Same issues of `examples/test_fec.py` also present in `examples/test_fec.ipynb`
- Error when running `ipython examples/test_NLC_withDBP_WDM_transmission.py`: `ImportError: cannot import name 'phaseNoise' from 'optic.models.channels`
- Large deviations between theory and Monte Carlo in `examples/test_metrics.ipynb` for small SNRs. What is the reason behind this? (I have also increased the number of samples by a factor of 10, which did not yield any improvement.)
- ImportError in `examples/test_ofdm.py`

## Examples
- ImportError in `examples/basic_EDFA.py`

## Paper (Minor Stuff)
- Acronym DSP is defined twice
- Non-breaking space `~` in "Fig.~3" is actually printed as a tilde.
