+++ docs/build/html/_sources/generated/optic.comm.modulation.pamConst.rst.txt
@@ -1,0 +1,6 @@
+﻿optic.comm.modulation.pamConst
+==============================
+
+.. currentmodule:: optic.comm.modulation
+
+.. autofunction:: pamConst
