+++ docs/requirements.txt
@@ -1,1 +1,2 @@
 opticommpy>=0.6.0
+nbsphinx
