### 2024-04-24 ca3800c
add references
