+++ optic/comm/modulation.py
@@ -75,7 +75,7 @@ def grayMapping(M, constType):
     
     References
     ----------
-    [1] Proakis, J. G., & Salehi, M. (2008). Digital Communications (5th Edition). McGraw-Hill Education.
+    [1] Proakis, J. G., & Salehi, M. Digital Communications (5th Edition). McGraw-Hill Education, 2008.
     """
     if M != 2 and constType == "ook":
         logg.warn("OOK has only 2 symbols, but M != 2. Changing M to 2.")
@@ -127,7 +127,7 @@ def pamConst(M):
     
     References
     ----------
-    [1] Proakis, J. G., & Salehi, M. (2008). Digital Communications (5th Edition). McGraw-Hill Education.
+    [1] Proakis, J. G., & Salehi, M. Digital Communications (5th Edition). McGraw-Hill Education, 2008.
     """
     L = int(M - 1)
     return np.arange(-L, L + 1, 2)
@@ -149,7 +149,7 @@ def qamConst(M):
 
     References
     ----------
-    [1] Proakis, J. G., & Salehi, M. (2008). Digital Communications (5th Edition). McGraw-Hill Education.
+    [1] Proakis, J. G., & Salehi, M. Digital Communications (5th Edition). McGraw-Hill Education, 2008.
     """
     L = int(np.sqrt(M) - 1)
 
@@ -183,7 +183,7 @@ def pskConst(M):
 
     References
     ----------
-    [1] Proakis, J. G., & Salehi, M. (2008). Digital Communications (5th Edition). McGraw-Hill Education.
+    [1] Proakis, J. G., & Salehi, M. Digital Communications (5th Edition). McGraw-Hill Education, 2008.
     """
     # generate complex M-PSK constellation
     pskPhases = np.arange(0, 2 * np.pi, 2 * np.pi / M)
@@ -208,9 +208,9 @@ def apskConst(M, m1=None, phaseOffset=None):
 
     References
     ----------
-    [1] Z. Liu, et al "APSK Constellation with Gray Mapping," IEEE Communications Letters, vol. 15, no. 12, pp. 1271-1273, December 2011
+    [1] Z. Liu, et al "APSK Constellation with Gray Mapping," IEEE Communications Letters, vol. 15, no. 12, pp. 1271-1273, 2011.
 
-    [2] Proakis, J. G., & Salehi, M. (2008). Digital Communications (5th Edition). McGraw-Hill Education.
+    [2] Proakis, J. G., & Salehi, M. Digital Communications (5th Edition). McGraw-Hill Education, 2008.
     """
     if m1 is None:
         if M == 16:
@@ -275,7 +275,7 @@ def minEuclid(symb, const):
 
     References
     ----------
-    [1] Proakis, J. G., & Salehi, M. (2008). Digital Communications (5th Edition). McGraw-Hill Education.
+    [1] Proakis, J. G., & Salehi, M. Digital Communications (5th Edition). McGraw-Hill Education, 2008.
 
     """
     ind = np.zeros(symb.shape, dtype=np.int64)
@@ -303,7 +303,7 @@ def demap(indSymb, bitMap):
     
     References
     ----------
-    [1] Proakis, J. G., & Salehi, M. (2008). Digital Communications (5th Edition). McGraw-Hill Education.
+    [1] Proakis, J. G., & Salehi, M. Digital Communications (5th Edition). McGraw-Hill Education, 2008.
 
     """
     M = bitMap.shape[0]
@@ -346,7 +346,7 @@ def detector(r, σ2, constSymb, px=None, rule='MAP'):
 
     References
     ----------
-    [1] Proakis, J. G., & Salehi, M. (2008). Digital Communications (5th Edition). McGraw-Hill Education.
[…]
