+++ optic/models/modelsGPU.py
@@ -68,6 +68,10 @@ def edfa(Ei, param):
     Eo : np.array
         Amplified noisy optical signal.
 
+    References
+    ----------
+    [1] R. -J. Essiambre,et al, "Capacity Limits of Optical Fiber Networks," in Journal of Lightwave Technology, vol. 28, no. 4, pp. 662-701, 2010, doi: 10.1109/JLT.2009.2039464.
+
     """
     try:
         Fs = param.Fs
@@ -144,6 +148,13 @@ def ssfm(Ei, param):
         Optical signal after nonlinear propagation.
     param : parameter object  (struct)
         Object with physical/simulation parameters used in the split-step alg.
+
+    References
+    ----------
+    [1] G. P. Agrawal, Nonlinear Fiber Optics, Elsevier Science, 2013. 
+
+    [2] O. V. Sinkin, R. Holzlöhner, J. Zweck, e C. R. Menyuk, “Optimization of the split-step Fourier method in modeling optical-fiber communications systems”, Journal of Lightwave Technology, vol. 21, nº 1, p. 61–68, jan. 2003, doi: 10.1109/JLT.2003.808628.
+
 
     """
     try:
