+++ README.rst
@@ -11,7 +11,7 @@
 **OptiCommPy** is a Python-based framework to simulate systems, subsystems, and 
 components of fiber optic communication systems, for educational and research purposes.
 
-|eye| |spectrum|
+|eye||spectrum|
 
 |DSP|
 
