+++ README.rst
@@ -1,13 +1,6 @@
-.. .. image:: ../../figures/logo_OptiCommpy.jpg
-..     :width: 500px  
-..     :align: center
 
-.. .. image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/logo_OptiCommPy.jpg
-..    :width: 500px
-..    :align: center
-
-|logo|
-
+|logo| 
+     
 **OptiCommPy** is a Python-based framework to simulate systems, subsystems, and 
 components of fiber optic communication systems, for educational and research purposes.
 
@@ -121,14 +114,13 @@ Joaquin Matres. (2023). edsonportosilva/OptiCommPy: v0.7.0-alpha
 .. |DOI| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.10304945.svg
    :target: https://doi.org/10.5281/zenodo.10304945
 .. |DSP| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/DSP.jpg
-   :width: 100%
-   :align: center
+             :width: 1000px
+             :align: center
 .. |logo| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/logo_OptiCommPy.jpg
-   :width: 50%
-   :align: center
+          :align: center
 .. |eye| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/eyeDisp.gif   
-   :width: 50%
+          :width: 500px
 .. |spectrum| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/40GOOK_spectrum.jpg
-   :width: 50%
+          :width: 500px
    
 
