+++ README.rst
@@ -69,12 +69,14 @@ Requirements/Dependencies
 -  python>=3.2
 -  numpy>=1.9.2
 -  scipy>=0.15.0
--  matplotlib>=1.4.3
+-  matplotlib>=3.7.0
 -  scikit-commpy>=0.7.0
 -  numba>=0.54.1
 -  tqdm>=4.64.1
 -  simple-pid>=1.0.1
 -  mpl-scatter-density>=0.7.0
+-  pandas>=2.0.0
+-  sphinx-rtd-theme>=1.2.2
 
 Installation
 ------------
