+++ README.rst
@@ -77,6 +77,7 @@ Requirements/Dependencies
 -  mpl-scatter-density>=0.7.0
 -  pandas>=2.0.0
 -  sphinx-rtd-theme>=1.2.2
+-  nbsphinx>=0.93
 
 Installation
 ------------
