+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 opticommpy>=0.6.0
 nbsphinx>=0.9.3
+nbsphinx-link>=1.3.0
