+++ requirements.txt
@@ -9,3 +9,4 @@ mpl-scatter-density>=0.7.0
 pandas>=2.0.0
 sphinx-rtd-theme>=1.2.2
 nbsphinx>=0.9.3
+nbsphinx-link>=1.3.0
