+++ setup.py
@@ -44,7 +44,8 @@ setup(
         "mpl-scatter-density>=0.7.0",
         "pandas>=2.0.0",
         "sphinx-rtd-theme>=1.2.2",
-        "nbsphinx>=0.9.3"
+        "nbsphinx>=0.9.3",
+        "nbsphinx-link>=1.3.0"
     ],
     #'package' package must contain files (see list above)
     # This dict maps the package name =to=> directories
