+++ README.rst
@@ -1,22 +1,20 @@
 
-|logo| 
-     
+.. image:: ../../figures/logo_OptiCommPy.jpg
+   :width: 75%  
+   :align: center
+
 **OptiCommPy** is a Python-based framework to simulate systems, subsystems, and 
 components of fiber optic communication systems, for educational and research purposes.
 
-|eye||spectrum|
-
-|DSP|
-
-.. .. image:: ../../figures/eyeDisp.gif
-..      :width: 45%      
-.. .. image:: ../../figures/40GOOK_spectrum.jpg
-..      :width: 45% 
+.. image:: ../../figures/eyeDisp.gif
+     :width: 45%      
+.. image:: ../../figures/40GOOK_spectrum.jpg
+     :width: 45% 
      
 
-.. .. image:: ../../figures/DSP.jpg
-..      :width: 600px  
-..      :align: center
+.. image:: ../../figures/DSP.jpg
+     :width: 600px  
+     :align: center
 
 |Binder| |Documentation Status| |DOI|
 
@@ -117,14 +115,14 @@ Joaquin Matres. (2023). edsonportosilva/OptiCommPy: v0.7.0-alpha
    :target: https://opticommpy.readthedocs.io/en/latest/?badge=latest
 .. |DOI| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.10304945.svg
    :target: https://doi.org/10.5281/zenodo.10304945
-.. |DSP| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/DSP.jpg
-             :width: 1000px
-             :align: center
-.. |logo| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/logo_OptiCommPy.jpg
-          :align: center
-.. |eye| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/eyeDisp.gif   
-          :width: 500px
-.. |spectrum| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/40GOOK_spectrum.jpg
-          :width: 500px
+.. .. |DSP| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/DSP.jpg
+..              :width: 1000px
+..              :align: center
+.. .. |logo| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/logo_OptiCommPy.jpg
+..           :align: center
+.. .. |eye| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/eyeDisp.gif   
+..           :width: 500px
+.. .. |spectrum| image:: https://github.com/edsonportosilva/OptiCommPy/blob/main/figures/40GOOK_spectrum.jpg
+..           :width: 500px
    
 
