+++ examples/simple_example.py
@@ -83,113 +83,114 @@ Pb = 0.5 * erfc(Q / np.sqrt(2))
 print(f"Pb = {Pb:.2e}\n")
 
 
-# # Generate BER vs Pin curve
+# Generate BER vs Pin curve
 
-# # simulation parameters
-# SpS = 16        # Samples per symbol
-# M = 2           # order of the modulation format
-# Rs = 10e9       # Symbol rate (for the OOK case, Rs = Rb)
-# Fs = SpS*Rs     # Signal sampling frequency (samples/second)
-# Ts = 1/Fs       # Sampling period
+# simulation parameters
+SpS = 16        # Samples per symbol
+M = 2           # order of the modulation format
+Rs = 10e9       # Symbol rate (for the OOK case, Rs = Rb)
+Fs = SpS*Rs     # Signal sampling frequency (samples/second)
+Ts = 1/Fs       # Sampling period
 
-# # MZM parameters
-# paramMZM = parameters()
-# paramMZM.Vpi = 2
-# paramMZM.Vb = -paramMZM.Vpi/2
+# MZM parameters
+paramMZM = parameters()
+paramMZM.Vpi = 2
+paramMZM.Vb = -paramMZM.Vpi/2
 
-# # typical NRZ pulse
-# pulse = pulseShape('nrz', SpS)
-# pulse = pulse/max(abs(pulse))
+# typical NRZ pulse
+pulse = pulseShape('nrz', SpS)
+pulse = pulse/max(abs(pulse))
 
-# # linear fiber optical channel parameters
-# paramCh = parameters()
-# paramCh.L = 100  # total link distance [km]
-# paramCh.alpha = 0.2  # fiber loss parameter [dB/km]
-# paramCh.D = 16  # fiber dispersion parameter [ps/nm/km]
-# paramCh.Fc = 193.1e12  # central optical frequency [Hz]
-# paramCh.Fs = Fs
+# linear fiber optical channel parameters
+paramCh = parameters()
+paramCh.L = 100  # total link distance [km]
+paramCh.alpha = 0.2  # fiber loss parameter [dB/km]
+paramCh.D = 16  # fiber dispersion parameter [ps/nm/km]
+paramCh.Fc = 193.1e12  # central optical frequency [Hz]
+paramCh.Fs = Fs
 
-# # photodiode parameters
-# paramPD = parameters()
-# paramPD.ideal = False
-# paramPD.B = Rs
-# paramPD.Fs = Fs
+# photodiode parameters
+paramPD = parameters()
+paramPD.ideal = False
+paramPD.B = Rs
+paramPD.Fs = Fs
 
-# powerValues = np.arange(-30, -9) # power values at the input of the pin receiver
-# distances = np.arange(100,160,10)
+powerValues = np.arange(-30, -9) # power values at the input of the pin receiver
+distances = np.arange(100,160,10)
 
-# BER = np.zeros((len(distances), len(powerValues)))
-# Q = np.zeros((len(distances), len(powerValues)))
+BER = np.zeros((len(distances), len(powerValues)))
+Q = np.zeros((len(distances), len(powerValues)))
 
-# for indL, L in enumerate(distances):
-#     paramCh.L = L  # total link distance [km]
-#     for indPi, Pi_dBm in enumerate(powerValues):        
-#         Pi = dBm2W(Pi_dBm + L*paramCh.alpha + 3) # optical signal power in W at the MZM input
+for indL, L in enumerate(distances):
+    paramCh.L = L  # total link distance [km]
+    for indPi, Pi_dBm in enumerate(powerValues):        
[…]
