### 2024-04-25 ad25ccc
remove .py files from examples folder
