+++ examples/test_metrics.py
@@ -1,563 +1,0 @@
-# -*- coding: utf-8 -*-
-# ---
-# jupyter:
-#   jupytext:
-#     formats: ipynb,py:light
-#     text_representation:
-#       extension: .py
-#       format_name: light
-#       format_version: '1.5'
-#       jupytext_version: 1.14.7
-#   kernelspec:
-#     display_name: Python 3 (ipykernel)
-#     language: python
-#     name: python3
-# ---
-
-# <a href="https://colab.research.google.com/github/edsonportosilva/OptiCommPy/blob/main/examples/test_metrics.ipynb" target="_parent"><img src="https://colab.research.google.com/assets/colab-badge.svg" alt="Open In Colab"/></a>
-
-if 'google.colab' in str(get_ipython()):    
-    # ! git clone -b main https://github.com/edsonportosilva/OptiCommPy
-    from os import chdir as cd
-    cd('/content/OptiCommPy/')
-    # ! pip install .
-
-from optic.comm.modulation import modulateGray, grayMapping
-from optic.comm.metrics import monteCarloGMI, monteCarloMI, fastBERcalc, theoryBER, theoryMI, calcEVM
-from optic.models.channels import awgn
-from optic.dsp.core import pnorm, signal_power
-from optic.plot import pconst
-from optic.utils import parameters, dB2lin
-import matplotlib.pyplot as plt
-import numpy as np
-import scipy as sp
-from tqdm.notebook import tqdm
-from numba import njit
-
-# %load_ext autoreload
-# %autoreload 2
-# #%load_ext line_profiler
-
-# ## Test bit-error-rate (BER) versus signal-to-noise ratio per bit ($E_b/N_0$)
-
-# ### QAM constellations with Gray mapping
-
-# +
-# Run BER vs Ebn0 Monte Carlo simulation in the AWGN channel
-
-qamOrder  = [4, 16, 64, 256, 1024, 4096]  # Modulation order
-
-EbN0dB_  = np.arange(-2, 35, 1)
-BER      = np.zeros((len(EbN0dB_),len(qamOrder)))
-BER[:]   = np.nan
-
-for ii, M in enumerate(qamOrder):
-    print('run sim: M = ', M)
-    for indSNR in tqdm(range(EbN0dB_.size)):
-
-        EbN0dB = EbN0dB_[indSNR]
-
-        # generate random bits
-        bitsTx = np.random.randint(2, size=int(np.log2(M)*1e5))    
-
-        # Map bits to constellation symbols
-        symbTx = modulateGray(bitsTx, M, 'qam')
-
-        # Normalize symbols energy to 1
-        symbTx = pnorm(symbTx) 
-
-        # AWGN channel  
-        snrdB  = EbN0dB + 10*np.log10(np.log2(M))
-        symbRx = awgn(symbTx, snrdB)
-
-        # BER calculation
-        BER[indSNR, ii], _, _ = fastBERcalc(symbRx, symbTx, M, 'qam')
-
-        if BER[indSNR, ii] == 0:              
-            break
-
-
-# +
-# Plot simulation results and theoretical curves        
-BER[BER==0] = np.nan
-
-plt.figure(figsize=(10,6))
-for ii, M in enumerate(qamOrder):
[…]
