+++ examples/OOK_transmission_with_advanced_EDFA_model.py
@@ -1,338 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     formats: ipynb,py:light
-#     text_representation:
-#       extension: .py
-#       format_name: light
-#       format_version: '1.5'
-#       jupytext_version: 1.15.1
-#   kernelspec:
-#     display_name: Python 3 (ipykernel)
-#     language: python
-#     name: python3
-# ---
-
-# <a href="https://colab.research.google.com/github/edsonportosilva/OptiCommPy/blob/main/jupyter/basic_OOK_transmission.ipynb" target="_parent"><img src="https://colab.research.google.com/assets/colab-badge.svg" alt="Open In Colab"/></a>
-
-# # Simulate a OOK transmission system with an advanced EDFA model
-
-if 'google.colab' in str(get_ipython()):    
-    # ! git clone -b main https://github.com/edsonportosilva/OptiCommPy
-    from os import chdir as cd
-    cd('/content/OptiCommPy/')
-    # ! pip install . 
-
-# +
-import numpy as np
-from commpy.utilities  import upsample
-from optic.models.devices import mzm, photodiode
-from optic.dsp.core import signal_power, firFilter, pulseShape
-from optic.utils import parameters, dBm2W
-from optic.plot import eyediagram
-import matplotlib.pyplot as plt
-from scipy.special import erfc
-
-from scipy.constants import c
-from optic.models.amplification import edfaSM,OSA
-
-import os.path as path
-# -
-
-import logging as logg
-logg.getLogger().setLevel(logg.INFO)
-logg.basicConfig(format='%(message)s')
-
-from IPython.core.display import HTML
-from IPython.core.pylabtools import figsize
-
-HTML("""
-<style>
-.output_png {
-    display: table-cell;
-    text-align: center;
-    vertical-align: middle;
-}
-</style>
-""")
-
-figsize(10, 3)
-
-# ### Intensity modulation (IM) with On-Off Keying (OOK)
-
-# +
-# simulation parameters
-SpS = 8
-Rs = 10e9 # Symbol rate (for OOK case Rs = Rb)
-Fs = SpS*Rs # Signal sampling frequency (samples/second)
-Ts = 1/Fs # Sampling period
-
-
-# Laser power
-Pi_dBm = -16        # laser optical power at the input of the MZM in dBm
-Pi = dBm2W(Pi_dBm) # convert from dBm to W
-
-# MZM parameters
-paramMZM = parameters()
-paramMZM.Vpi = 2
-paramMZM.Vb = -paramMZM.Vpi/2
-
-# generate pseudo-random bit sequence
-bitsTx = np.random.randint(2, size=100000)
-n = np.arange(0, bitsTx.size)
-
-# map bits to electrical pulses
-symbTx = 2*bitsTx-1
-symbTx = symbTx/np.sqrt(signal_power(symbTx))
-
-# upsampling
-symbolsUp = upsample(symbTx, SpS)
-
-# typical NRZ pulse
-pulse = pulseShape('nrz', SpS)
-pulse = pulse/max(abs(pulse))
-
-# pulse formatting
-sigTx = firFilter(pulse, symbolsUp)
-
[…]

[…]
