+++ README.rst
@@ -76,7 +76,8 @@ Requirements/Dependencies
 -  pandas>=2.0.0
 -  sphinx-rtd-theme>=1.2.2
 -  nbsphinx>=0.9.3
--  nbsphinx-link>=1.3.0  
+-  nbsphinx-link>=1.3.0
+-  cupy-cuda12x >= 13.1.0 (optional, in GPU processing is desired)
 
 Installation
 ------------
