+++ README.rst
@@ -77,7 +77,7 @@ Requirements/Dependencies
 -  sphinx-rtd-theme>=1.2.2
 -  nbsphinx>=0.9.3
 -  nbsphinx-link>=1.3.0
--  cupy-cuda12x >= 13.1.0 (optional, in GPU processing is desired)
+-  cupy-cuda12x >= 13.1.0 (optional, in case GPU processing is desired)
 
 Installation
 ------------
