+++ README.md
@@ -55,6 +55,7 @@ If you want to contribute to this project, implement the feature you want and se
 -  sphinx-rtd-theme>=1.2.2
 -  nbsphinx>=0.9.3
 -  nbsphinx-link>=1.3.0
+-  cupy-cuda12x >= 13.1.0 (optional, in case GPU processing is desired)
 
 ## Installation
 
