### 2024-04-25 #6600 @edsonportosilva
Hello @lucydot @ebezzam. Sorry for my late reply. I am currently overloaded with the end of semester tasks. @ebezzam, thanks again for the time reviewing this project. Here is the reply to your queries.

> @lucydot, @edsonportosilva below is my initial review. I realize that there is an [examples folder](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples), but is it expected to run everything or is there a notebook/script that gives an overview of the main features?
>
> ### Functionality
> * **Installation**: I would recommend putting the same installation in your README and in the documentation. You can create a `README.rst` file, and use that within your documentation (like [this](https://github.com/ebezzam/python-dev-tips/blob/a76a766f506d871ad8e3a9f396ebb5bdc268addc/docs/source/index.rst?plain=1#L10) instead of creating a [separate file](https://github.com/edsonportosilva/OptiCommPy/blob/main/docs/source/usage.rst?plain=1) so you don’t have to edit twice). I say this because I looked at the [documentation](https://opticommpy.readthedocs.io/en/latest/usage.html) for the installation steps; and `python setup.py install` didn’t work for me, but `pip install .` did (which was in [your README](https://github.com/edsonportosilva/OptiCommPy/blob/452409d2ca8b604ef04aaffe675640eac21c628b/README.md?plain=1#L75)).

1. Fixed. Now the [documentation page](https://opticommpy.readthedocs.io/en/latest/index.html) should be updated as @ebezzam suggested. Moreover, we removed `python setup.py install` which is deprecated.

> * **Functionality**: it would be useful to have a section or notebook to show the [different features](https://opticommpy.readthedocs.io/en/latest/welcome.html), as I’m not sure where I would start to try things out. I realize there’s an [examples folder](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples), but it is not clear where to start.

2. Fixed. We have added to the documentation page a [getting started], as suggested by @lucydot, (https://opticommpy.readthedocs.io/en/latest/JOSS_paper_example.html) notebook where the reader can follow all the steps to set up a basic simulation and visualize performance results. This notebook also reproduces the plots shown in the JOSS paper. Another notebook was included with an advanced example, but I am still working to finish it.

> * **Performance**: it would be interesting to quantify the speed-up thanks to GPU acceleration, as mentioned in Lines 61-64 of the paper.

3. We have added to the examples folder a [notebook](https://github.com/edsonportosilva/OptiCommPy/blob/add-paper/examples/benchmarck_GPU_processing.ipynb) with a few basic benchmark results of CPU vs GPU processing time for simulations involving split-step Fourier algorithms, which are the most demanding processing tasks that could be performed in simulations . A mention of those results was included in the paper.

> ### Documentation
> * **A statement of need**: there is none in the [documentation](https://opticommpy.readthedocs.io/en/latest/index.html).

4. Fixed by 1.

> * **Automated tests**: there are none nor manual steps to verify functionality.

5. Partially fixed. We have included a few automated tests. However, the scope of OptiCommPy (communications, optics, DSP) makes it difficult to set up meaningful unitary tests for all functions. For example, there are several algorithms of equalization and carrier recovery implemented, and each has particular requirements for correct operation, which depends on the specific scenario where they are tested. For those working in the field, this is well understood, though. Our first idea to workaround this issue was to add notebooks in the examples folder to work as a "systemic test" for the relevant functions, such that the user could see cases of use of a given function and then verify that the code was working properly. In practice, most of the code in OptiCommPy is validated with data from optical experiments, i.e. we have been able to use OptiCommPy to process data obtained from fiber optic transmission experiments and then verify that the code (e.g. physical models for the fiber, DSP algorithms, etc) work with real data. We could try to upload a few experimental traces to the repo and add a notebook to demonstrate real data processing, but the files to be uploaded may be large, and the experimental scenarios will not cover all functions. In any case, our plan is to include in the future as many automated tests as possible.

> * **Community guidelines**: there is in the [README](https://github.com/edsonportosilva/OptiCommPy?tab=readme-ov-file#how-can-i-contribute), but as mentioned above for “Installation”, it could be nice to use the README in your documentation, as I didn’t see the contribution guidelines at first.

6. Fixed by 1.
> ### Software paper
> * **A statement of need**: The “Statement of need” starts great by giving the relevant context, but I feel like there’s a paragraph missing after Line 40 to explicitly say something like “OptiCommPy addresses this need by…” and who the target audience is (similar to what’s said in  “Summary”).

7. Fixed

> * **Quality of writing**: I think Lines 41 and onward should be in a separate section, e.g. with the title “OptiCommPy Code Structure”. And another section on what example scripts/code for the different features could be useful.

8. Fixed.

> ### Miscellaneous
> * Adding [.ipynb_checkpoints](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples/.ipynb_checkpoints) and [.pylint.d](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples/.pylint.d) to your `.gitignore`

9. Fixed.
