### 2024-04-26 #6600 @edsonportosilva
@taladjidi, thanks for your time reviewing this project. Please, see below the replies to the points you listed.

> @lucydot @ebezzam
>
> Sorry for the late reply, I was in leave. Below are my first comments.
>
> # Functionality
> * Installation went without trouble, though I only tried through `pip install .`
> * Examples: some of the notebooks do not run (while I did not test them all, [`basic_EDFA.ipynb`](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/basic_EDFA.ipynb) seems to use an old package structure and does not run. I ran this first since I use an EDFA in the lab and it was the first example in the directory.

1. Fixed. The examples folder has been updated with some unused files removed and all the remaining notebooks should be running now without issues. Let me know if there is still something broken.

> * Performance: I agree with @ebezzam that benchmarks curves would be a welcome addition. This could maybe be one of the tests ?

2. Fixed, partially. We have included a [notebook](https://github.com/edsonportosilva/OptiCommPy/blob/add-paper/examples/benchmarck_GPU_processing.ipynb) in the examples folder with a benchmark for the most critical GPU processing tasks that may be required in the simulations available in the repository, but not covering the whole package. However, we are not claiming any particular performance gain when compared to any other simulator. We have just stated that GPU processing may speedup the simulations when compared to the code without it, and the results in the benchmark in the examples folder now provides evidence that supports this claim.

>   Furthermore,  you will gain at least a factor of 2 performance on your FFT's if you explicitely define and cache the FFT plans (see [`get_fft_plan`](https://docs.cupy.dev/en/stable/reference/generated/cupyx.scipy.fftpack.get_fft_plan.html#cupyx.scipy.fftpack.get_fft_plan) or directly through `cupy.cuda.cufft`) instead of using the `cupyx.scipy.fft` API.
> * Performance / FFT: I would advise switching to a more performant FFT library than Numpy's since most of you DSP will rely on it. There are a lot of potential candidates, pyFFTW could be a drop in replacement. In the same vein, a lot of operations could be parallelized and JIT compiled using Numba.

3. Thanks for bringing up this point. Indeed, we have spent some time trying to optimize the performance of the split-step algorithms running with CuPy. I believe currently the implementation of FFT plans is done automatically by `cupyx.scipy.fft`. If you have any particular insights on how to further optimize performance, I would be glad to hear more.

> # Documentation
> * Examples: Jupyter notebooks are a great way to introduce the various functionalities of your software. I found that the graphical outputs were very clear. Maybe it would be good to add Markdown blocks in between the code blocks to link the code with the physical models in use (how does one equation translate in terms of your package functionalities).

4. Thanks for the observation. Yes, that is our idea in the long run. However, what we observe is that this is not a big issue for the public that has an interest in using OptiCommPy, which is usually someone with a little bit of experience in the field. As you noted, the graphical outputs are sometimes more important for them. However, since we and other colleagues abroad are using OptiCommPy when teaching courses, having more information on the examples is more important.

> * Physical models: I would expect the docstrings of the models to cite some relevant paper (when needed) explaining the model more in detail, or one or two latex equations to succintly define the used conventions etc ...

5. Fixed. We have included in the docstrings in most parts of the code citations to papers/books where one can find physical models, algorithms, equations, etc. described in detail.

> * In the paper, you compare to commercial / proprietary solutions, maybe you could include a section of your Readme to explain what funcitonalities are supported, or some kind of migration guide ? In a similar fashion, maybe you could include some performance comparisons.

6. Indeed, that would be interesting. However, those commercial simulators are expensive and we do not have the licenses necessary to make a comparison with OptiCommPy. I didn't get the part of the "migration guide" that you have mentioned.

> # Software paper
> * The manuscript overall is well written and has a very good introduction.
> * It would be better to split the Statement of need and the functionalities that you introduce, in order to highlight the problem on one side and its solution on the other.
> * The package structure is maybe superfluous in the paper, and maybe could be replaced with benchmarks / comparisons with proprietary software (if you have access to it) ?
>
> Overall it's a very nice project, while the examples need a bit of documentation work, they do seem to cover most of the functionalities of the package.

Thanks! :)

>
> Best,
