### 2024-04-26 #6600 @edsonportosilva
Hello @lucydot @klb2 thanks for your time reviewing this project. Please, see below the replies to the points you listed.

> @lucydot @edsonportosilva
>
> I have now completed my review. I tried to limit my comments to new ones that have not been covered in the reviews above.
>
> In general, the package is interesting and contains a lot of features. The main issues are related to the tests and documentation, which lacks significant information (even missing whole modules).
>
> ## Documentation
> * It would be nice to add a simple example after the installation instructions to verify that the installation worked.

1 . Fixed, partially. The content in the [getting started](https://opticommpy.readthedocs.io/en/latest/JOSS_paper_example.html) notebook available in the documentation should run without issues.

> * In general, the documentation does not show any usage examples.

2. Fixed. We have added to the docs two examples, a basic and an advanced.

> * Required version of matplotlib in documentation (>=1.4.3) does not match that in requirements/setup file (>=3.7.0)

3. Fixed

> * Pandas is not stated as a dependency in documentation

4. Fixed

> * It should be mentioned which scripts need to be run to reproduce the figures from the paper.

5. Fixed. The [getting started](https://opticommpy.readthedocs.io/en/latest/JOSS_paper_example.html) notebook reproduces the plots from the paper.

> * The documentation of the `fec` module is missing in the documentation.

6. Fixed. The FEC module has been removed because it was just unfinished work that has been discontinued.

> * Community guidelines, e.g., a `CONTRIBUTING.md` file, are missing

7. Fixed. The documentation now provides guidelines for contributing to the code.

> ## Tests
> * `tests/test_dsp.py` does not run with the `unittest` package

8. Could you elaborate on that issue? I can see that all tests in the test folder run with pytest.

> * The setup script (`setup.py`) says that the `nose` module is required for testing, but it is not actually required to run the unit tests in `tests/`

9. Fixed.

> * `examples/test_*.py` can not be run with Python (ipython is required). However, this is not mentioned anywhere.

10. Fixed. The scripts '.py' in the examples folder where redundant copies of the notebooks made by jupytext and were removed to avoid confusion with actual .py tests files.

> * There only seems to be a small number of unit tests. The code coverage is very low.

11. Yes, you are right. We are working towards increasing the coverage in the long run, but for reasons explained before this is not an easy task given the scope of the repository. However, the notebooks in the examples folder should be enough to verify the reproducibility and correctness for the main features of the repository.

> * The eye diagrams in `examples/test_dsp_core_functions.ipynb` are not useful (just a blue box).

12. Fixed.

> * Error when running `ipython examples/test_clockRecovery.py`: `ValueError: operands could not be broadcast together with shapes (44000,) (43968,)` (Line 121)

13. Redundant file removed.

> * Error when running `ipython examples/test_fec.py`: `ModuleNotFoundError: No module named 'optic.modulation'` (I think, it should be `optic.comm.modulation`)

14. File removed.

> * Also, the `optic.fec` module does not exist. I could not find any LDPC functionality in the documentation.

15. Fixed in 6.

> * Same issues of `examples/test_fec.py` also present in `examples/test_fec.ipynb`

16. Fixed in 6.

> * Error when running `ipython examples/test_NLC_withDBP_WDM_transmission.py`: `ImportError: cannot import name 'phaseNoise' from 'optic.models.channels`

17. Fixed.

> * Large deviations between theory and Monte Carlo in `examples/test_metrics.ipynb` for small SNRs. What is the reason behind this? (I have also increased the number of samples by a factor of 10, which did not yield any improvement.)

18. Those results are correct because the performance comparison was between APSK and PSK, and since APSK tends to perform better than PSK, the curves make sense.

> * ImportError in `examples/test_ofdm.py`

19. File removed.

> ## Examples
> * ImportError in `examples/basic_EDFA.py`

20. File removed.

> ## Paper (Minor Stuff)
> * Acronym DSP is defined twice

21. Fixed.

> * Non-breaking space `~` in "Fig.~3" is actually printed as a tilde.

22. Fixed.
