+++ docs/build/html/_sources/getting_started.nblink.txt
@@ -1,0 +1,3 @@
+{
+  "path": "../../examples/getting_started.ipynb"
+}
