### 2024-05-07 #6600 @klb2
Thank you for addressing my comments @edsonportosilva. I checked the code and documentation again and only some issues are remaining.

- The file `Image("./figures/BasicIMDD.png", width=600)` in the [`getting_started.ipynb`](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/getting_started.ipynb) notebook does not exist.
- I am still not sure about the correctness of the results in the [`test_metrics.ipynb`](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/test_metrics.ipynb) notebook.
  - First, for low SNRs, the Monte Carlo simulation results systematically deviate from the theoretical results (Sections 1.1 and 1.2). This should not be the case, especially for small error rates with enough samples.  ![2024-05-07-124244_837x519_scrot](https://github.com/openjournals/joss-reviews/assets/58686048/1742f479-24be-4222-9553-e1068a76397b)
  - Next, in Section 2.1 of the notebook, the 256 QAM theory curve breaks the Shannon limit. This appears to be wrong, especially since the Monte Carlo simulations again show the expected behavior.  ![Uploading 2024-05-07-124257_850x517_scrot.png…]()
- I did not see this before, but it should be "Fig. 2" instead of "Fig. 3" in the paper (line 82 on page 3).

Regarding point 8 from the previous comments: In the original submission, it was not clear/documented that pytest would be required to run the unit tests. Only the `unittest` module was imported in two out of the three files. But I saw that this is fixed now.
