+++ requirements.txt
@@ -1,7 +1,6 @@
 numpy>=1.9.2
 scipy>=0.15.0
 matplotlib>=3.7.0
-scikit-commpy>=0.7.0
 numba>=0.54.1
 tqdm>=4.64.1
 simple-pid>=1.0.1
