+++ setup.py
@@ -33,13 +33,12 @@ setup(
     packages=["optic", "optic.comm", "optic.models", "optic.dsp"],
     # py_modules = ['plot', 'core', 'comm', 'dsp', 'models'],
     install_requires=[
-        "numpy>=1.9.2",
-        "scipy>=0.15.0",
+        "numpy>=1.24.4",
+        "scipy>=1.13.0",
         "matplotlib>=3.7.0",
         "sympy",
         "tqdm>=4.64.1",
-        "numba>=0.54.1",
-        "scikit-commpy>=0.7.0",
+        "numba>=0.54.0,<=0.57",
         "simple-pid>=1.0.1",
         "mpl-scatter-density>=0.7.0",
         "pandas>=2.0.0",
