+++ README.rst
@@ -65,11 +65,10 @@ Requirements/Dependencies
 -------------------------
 
 -  python>=3.2
--  numpy>=1.9.2
--  scipy>=0.15.0
+-  numpy>=1.24.4
+-  scipy>=1.13.0
 -  matplotlib>=3.7.0
--  scikit-commpy>=0.7.0
--  numba>=0.54.1
+-  numba>=0.54.1,<=0.57.0
 -  tqdm>=4.64.1
 -  simple-pid>=1.0.1
 -  mpl-scatter-density>=0.7.0
