+++ requirements.txt
@@ -5,7 +5,6 @@ numba>=0.57.0
 tqdm>=4.64.1
 simple-pid>=1.0.1
 mpl-scatter-density>=0.7.0
-pandas>=2.0.0
 sphinx-rtd-theme>=1.2.2
 nbsphinx>=0.9.3
 nbsphinx-link>=1.3.0
