+++ requirements.txt
@@ -1,7 +1,7 @@
 numpy>=1.24.4
 scipy>=1.13.0
 matplotlib>=3.7.0
-numba>=0.57.0
+numba>=0.54.0,<=0.57.0
 tqdm>=4.64.1
 simple-pid>=1.0.1
 mpl-scatter-density>=0.7.0
