### 2024-05-09 #6600 @edsonportosilva
Hello @klb2, thanks again for the feedback,

> Thank you for addressing my comments @edsonportosilva. I checked the code and documentation again and only some issues are remaining.
>
> * The file `Image("./figures/BasicIMDD.png", width=600)` in the [`getting_started.ipynb`](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/getting_started.ipynb) notebook does not exist.

1. I don't know the problem in this case, because for me the figure is displayed both in the repository and the documentation page. Does anyone else see the same issue? @lucydot @ebezzam @taladjidi

> * I am still not sure about the correctness of the results in the [`test_metrics.ipynb`](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/test_metrics.ipynb) notebook.
>
>   * First, for low SNRs, the Monte Carlo simulation results systematically deviate from the theoretical results (Sections 1.1 and 1.2). This should not be the case, especially for small error rates with enough samples.  ![2024-05-07-124244_837x519_scrot](https://private-user-images.githubusercontent.com/58686048/328491842-1742f479-24be-4222-9553-e1068a76397b.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MTUxOTcxNDEsIm5iZiI6MTcxNTE5Njg0MSwicGF0aCI6Ii81ODY4NjA0OC8zMjg0OTE4NDItMTc0MmY0NzktMjRiZS00MjIyLTk1NTMtZTEwNjhhNzYzOTdiLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA1MDglMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwNTA4VDE5MzQwMVomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTk0NjQ5Mjk1NjE1Yjg4MDJiZTVmYzE4OTAwNWU0MGRhY2ViMWM0NzFmNTU5YmI1YjI1MzQyMzJjMDQ2NzQ2M2QmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.QalcJYucsrAheRdzbID0QI2L35HEOa4lLtlpXKIKoo4)

2. The deviation from Monte Carlo to theoretical results is expected in this case for large QAM/PSK constellations. This happens because the expression used to calculate the error probability assumes that the error events are dominated by errors between neighboring QAM/PSK symbols, in a gray-mapped constellation. That is, every symbol error translates into a bit error. However, this is only true in moderate to low symbol-error-rate (SER) regimes. If the SER is above 1e-1, this assumption is not true anymore and the Monte Carlo results should deviate from the approximated theoretical curves, which would then underestimate the BER. I have added an explanation to the function's ([theoryBER](https://opticommpy.readthedocs.io/en/latest/generated/optic.comm.metrics.theoryBER.html)) docstring to clarify this point. Nevertheless, such regimes of high error rates are mostly of no practical relevance for standard digital communication systems, including fiber optic communications.

>   * Next, in Section 2.1 of the notebook, the 256 QAM theory curve breaks the Shannon limit. This appears to be wrong, especially since the Monte Carlo simulations again show the expected behavior, e.g., see Fig. 1(a) in [D. S. Millar, T. Fehenberger, T. Koike-Akino, K. Kojima and K. Parsons, "Coded Modulation for Next-Generation Optical Communications," 2018 Optical Fiber Communications Conference and Exposition (OFC), 2018](http://dx.doi.org/10.1364/OFC.2018.Tu3C.3).
>     ![2024-05-07-124257_850x517_scrot](https://private-user-images.githubusercontent.com/58686048/328494704-f50507a2-445f-4ade-b849-7638f6698869.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MTUxOTcxNDEsIm5iZiI6MTcxNTE5Njg0MSwicGF0aCI6Ii81ODY4NjA0OC8zMjg0OTQ3MDQtZjUwNTA3YTItNDQ1Zi00YWRlLWI4NDktNzYzOGY2Njk4ODY5LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA1MDglMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwNTA4VDE5MzQwMVomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTkyMjQzYWVkYmYzYzgxYzgxNjk3MGY0MDBhZGFhOThhMTlkODQ0YThlZGJkMjE1N2JmYWUwMTc2MzJiZDlmMjEmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.9AIVE0NQJzJ_E9sxS0gdCZZn0wKKgFEhVEksfUxyNyk)

3. Thanks for reporting this issue. This deviation had not appeared before and I did a quick investigation to find that is related to the version of Numba. Results are fine with versions 0.54.0-0.57.0 and starting from version 0.58.0 this deviation appears. These results are obtained by integrating numerically the theoretical expression of the mutual information calculated between input and output for the memoryless Gaussian channel. It requires a double integral to be calculated with [dblquad](https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.dblquad.html#scipy.integrate.dblquad), whose argument is compiled with Numba to speed up the calculations. It turns out that different versions of Numba appear to be giving different precision for the calculation of this function, and some numerical instability is causing the results to deviate from the correct curves. Due to the lack of a better solution, for now, I have restricted the version of Numba in the requirements to 0.54.0-0.57.0 to work around this issue. See the results below.

![image](https://github.com/openjournals/joss-reviews/assets/17885240/d1138761-c5cb-4697-b7d4-3b0e3c39623f)

> * I did not see this before, but it should be "Fig. 2" instead of "Fig. 3" in the paper (line 82 on page 3).

4. Fixed.

> Regarding point 8 from the previous comments: In the original submission, it was not clear/documented that pytest would be required to run the unit tests. Only the `unittest` module was imported in two out of the three files. But I saw that this is fixed now.
