### 2024-05-13 #6600 @lucydot
> I don't know the problem in this case, because for me the figure is displayed both in the repository and the documentation page. Does anyone else see the same issue?

It is displayed here: https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/getting_started.ipynb. But when I look in the figures folder, it is true the image does not exist. So I think if you run locally, it will give the reported error. The notebook being displayed online must have been executed before the image was removed. In fact, the path to the figures folder also seems to have changed.

@ebezzam @taladjidi - a reminder to re-visit your checklist when you get a chance, and see if there are any outstanding issues blocking acceptance.
