### 2024-05-13 #6600 @taladjidi
Hi @lucydot , on my side it looks good ! Thank you @ebezzam for the swift reply, I will try to draft a PR for the FFT / kernel performance stuff I identified.
