### 2024-05-13 #6600 @klb2
Thanks for addressing my remaining comments @edsonportosilva.

One last suggestion is that you could add (and emphasize) on the main page of the documentation that the software is open-source.
However, I guess that is optional and I have completed my review @lucydot.
