### 2024-05-15 #6600 @edsonportosilva
Thanks for the feedback, @ebezzam @lucydot!

> @edsonportosilva thank you for addressing my comments. The [Getting started](https://opticommpy.readthedocs.io/en/latest/getting_started.html) page with interleaved code and description looks very nice.

Thanks :)

> Just some minor stuff, but otherwise it looks good for me:
>
> * The "Open in Colab" button doesn't work for me. From [ReadTheDocs](https://opticommpy.readthedocs.io/en/latest/getting_started.html), it is not clickable and from [GitHub](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/getting_started.ipynb) I get the error `Notebook not found: Could not find JOSS_paper_example.ipynb in https://api.github.com/repos/edsonportosilva/OptiCommPy/contents/examples?`

1. Should be fixed by now in the GitHub notebook. On the documentation page, for some reason, the logo is not clickable, and I noticed that happening for a few other badges. I assume this is related to some bug in the translation of the notebook content to an RST document. I will try to find a solution for that but, for now, the link in the repository should be enough to open the notebook in Colab.

> * In your paper, it could be nice to have an "Examples" section with links, e.g. adapting lines 83-87 as:
>
> ```
> # Examples
>
> In the documentation, one can find [an example](https://opticommpy.readthedocs.io/en/latest/getting_started.html) that demonstrates core features of OptiCommPy. A collection of examples to build several different simulation setups, including advanced setups with non-linear fiber propagation models, WDM transmission, and coherent detection can be found in the repository's [examples](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples) folder. [Benchmarks](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/benchmarck_GPU_processing.ipynb) quantifying the speedup achieved by using GPU acceleration are also provided.
> ```
>
> which renders as
>
> # Examples
> In the documentation, one can find [an example](https://opticommpy.readthedocs.io/en/latest/getting_started.html) that demonstrates core features of OptiCommPy. A collection of examples to build several different simulation setups, including advanced setups with non-linear fiber propagation models, WDM transmission, and coherent detection can be found in the repository's [examples](https://github.com/edsonportosilva/OptiCommPy/tree/main/examples) folder. [Benchmarks](https://github.com/edsonportosilva/OptiCommPy/blob/main/examples/benchmarck_GPU_processing.ipynb) quantifying the speedup achieved by using GPU acceleration are also provided.

2. Thanks for the suggestion and we have changed the text accordingly.
