### 2024-05-20 #6600 @lucydot
@edsonportosilva, please adjust the paper so that when you have multiple citations are contained in the same bracket. This can be done using semi-colons to separate items (see [here](https://pandoc.org/MANUAL.html#extension-citations)). After you have adjusted the paper, please re-compile using `@editorialbot generate pdf` to check all are formatted as expected.
