### 2024-06-03 2e6bf9d
add orcid, fix citations
