### 2024-06-03 #6600 @edsonportosilva
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete

Hello, @lucydot! Sorry for the late reply, last two weeks I have been unable to proceed with the final steps of the review. Here it is:

> * Double check authors and affiliations (including ORCIDs)

1. Done

> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

2. https://github.com/edsonportosilva/OptiCommPy/releases/tag/v0.9.0-alpha

> * Archive the release on Zenodo/figshare/etc and post the DOI here.

3.  https://doi.org/10.5281/zenodo.11450597

> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

4. Done.

> * Make sure that the license listed for the archive is the same as the software license.

5. Done
