### 2024-06-05 #6600 @lucydot
Hi @edsonportosilva, can you check and confirm that there are no DOIs available for the references above? I can see they are either software or book references, so less likely to be available than for journal references.

After that I will recommend acceptance to the track editors 🥳
