### 2024-06-05 #6600 @edsonportosilva
Hi @lucydot,

> Hi @edsonportosilva, can you check and confirm that there are no DOIs available for the references above? I can see they are either software or book references, so less likely to be available than for journal references.

I have checked and there is no DOI available for those references, mostly because they are either software or physical books, as you noted.

> After that I will recommend acceptance to the track editors 🥳

Great to hear that and thank you for your guidance during this peer review process 😊
