### 2024-06-06 280989a
remove redundant syntax
