+++ paper/paper.bib
@@ -3,8 +3,6 @@
   author = {Agrawal, Gowind P.},
   biburl = {https://www.bibsonomy.org/bibtex/268655f428f3d12636ba631a76eea8d9b/kaqmak},
   edition = {3rd},
-  interhash = {364c70c7807f1e98b2521561352efdb8},
-  intrahash = {68655f428f3d12636ba631a76eea8d9b},
   keywords = {},
   publisher = {John Wiley \& Sons},
   timestamp = {2010-02-16T09:19:52.000+0100},
@@ -16,8 +14,6 @@
 abstract = {We report here on our investigations of the Manakov-polarization mode dispersion (PMD) equation which can be used to model both nonreturn-to-zero (NRZ) and soliton signal propagation in optical fibers with randomly varying birefringence. We review the derivation of the Manakov-PMD equation from the coupled nonlinear Schr{\"{o}}dinger equation, and We discuss the physical meaning of its terms. We discuss our numerical approach for solving this equation, and we apply this approach to both NRZ and soliton propagation. We show by comparison with the coupled nonlinear Schr{\"{o}}dinger equation, integrated with steps that are short enough to follow the detailed polarization evolution, that our approach is orders of magnitude faster with no loss of accuracy. Finally, we compare our approach to the widely used coarse-step method and demonstrate that the coarse-step method is both efficient and valid.},
 author = {Marcuse, Dietrich and Menyuk, C. R. and Wai, P. K.A.},
 doi = {10.1109/50.622902},
-isbn = {0733-8724},
-issn = {07338724},
 journal = {J. Light. Technol.},
 keywords = {Birefringence,Optical fibers,Polarizations,Random bifringence,Solitons},
 number = {9},
@@ -39,8 +35,6 @@ year = {1997}
 abstract = {Digital coherent receivers have caused a revolution in the design of optical transmission systems, due to the subsystems and algorithms embedded within such a receiver. After giving a high-level overview of the subsystems, the optical front end, the analog-to-digital converter (ADC) and the digital signal processing (DSP) algorithms, which relax the tolerances on these subsystems are discussed. Attention is then turned to the compensation of transmission impairments, both static and dynamic. The discussion of dynamic-channel equalization, which forms a significant part of the paper, includes a theoretical analysis of the dual-polarization constant modulus algorithm, where the control surfaces several different equalizer algorithms are derived, including the constant modulus, decision-directed, trained, and the radially directed equalizer for both polarization division multiplexed quadriphase shift keyed (PDM-QPSK) and 16 level quadrature amplitude modulation (PDM-16-QAM). Synchronization algorithms employed to recover the timing and carrier phase information are then examined, after which the data may be recovered. The paper concludes with a discussion of the challenges for future coherent optical transmission systems. {\textcopyright} 2010 IEEE.},
 author = {Savory, Seb J.},
 doi = {10.1109/JSTQE.2010.2044751},
-isbn = {1077-260X},
-issn = {1077260X},
 journal = {IEEE J. Sel. Top. Quantum Electron.},
 keywords = {Digital communication,polarization},
 month = {sep},
@@ -55,8 +49,6 @@ year = {2010}
[…]
