### 2024-06-10 #6600 @edsonportosilva
Hello @lucydot, I guess something was wrong with the ISBN information in the .bib file added to the paper. I have removed this information from the file since it is not displayed in the paper. Now, perhaps the process will run without errors.
