### 2024-06-11 #6600 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
