### 2024-06-11 #6600 @kyleniemeyer
Congratulations @edsonportosilva on your article's publication in JOSS! rePlease consider [signing up as a reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.
